vencrypty.py
Arguments - The program should take three command-line arguments. The first argument is the filename of a file containing the plaintext. The second argument is the name of a file where the ciphertext will be written. The third argument is a key to be used for encryption. 

Output - The program should create a file, with the appropriate name, containing the ciphertext. The program should print nothing to the screen.