vdecrypt.py
Arguments - The program should take three command-line arguments. The first argument is the filename of a file containing the ciphertext. The second argument is the name of a file where the plaintext will be written. The third argument is a key to be used for decryption.

Output - The program should create a file, with the appropriate name, containing the plaintext. The program should print nothing to the screen.