server.py
User Interface of the Server

 
The server should be started before the client. The server program should take three command-line arguments: 

Server Port - The number of the port which the server is listening to. 
Game Data File - The name of the file containing the database of all information about all games.
Private Key File - The name of the file containing the server’s private key. 
 

The Private Key File is a file whose only contents are the server’s private key. 

 

When the server is started, it should print “server started” and then wait to receive messages from a client. Only one client can connect to the server at a time. When the server receives a request from a client, the server will print the game_id and field in the message on the screen on a new line. An example of the printed output of the server when communicating with the client is shown below. 

server started

2018090600 game_id

2018090600 home_team

2018090600 away_team

2018090600 week

2018090600 season

2018090600 home_score

2018090600 away_score




The server will continue responding to requests from a client and printing the associated information until a quit message is received. At that point, the server will stop communicating with the client and wait for a new client to initiate communication. The server will continue to run in an infinite loop until its process is killed externally, for instance by a ctrl-c typed at the keyboard.