client.py
The client program should take three command-line arguments:

Server Host - This is the hostname on which the server program is running.
Server Port - This is the port number on the server host to which the server is listening to.
Server Public Key File - This is the name of the file which contains the server’s public key.
 

The Server Public Key File is a file whose only contents are the server’s public key.

 

The client will present the user with a prompt, “>”. The user makes requests for game information by entering game_id and field info at the client’s prompt. When this information is entered at the client’s prompt, the client will send a data request message containing the game_id and field to the server, and the client will await a data response message from the server. The server will send a data response message containing the corresponding information of the field for the match with that game_id. The client will print the received information to the screen, print a prompt on a new line, and wait for more input from the user. An example of a user interaction with the client is shown below. 




> 2018090600 type

reg

> 2018090600 game_id

2018090600

> 2018090600 home_team

PHI

> 2018090600 away_team

ATL

> 2018090600 week

1

> 2018090600 season

2018

> 2018090600 home_score

18

> 2018090600 away_score

12

> whats up?

unknown

> 2018090600 away_score

12

> quit


The client will continue in this loop until the user enters “quit”, at which point the client will send a quit message to the server and the client’s process will exit.